% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{assetMuSd}
\alias{assetMuSd}
\title{Calculate Portfolio Asset Mean and Standard Deviation}
\usage{
assetMuSd(logL)
}
\arguments{
\item{logL}{List of time series returns for each asset}
}
\value{
Mean and SD for each asset in logL
}
\description{
Calculate Portfolio Asset Mean and Standard Deviation
}
