% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{idxList}
\alias{idxList}
\title{Index column(s) in quandlList object}
\usage{
idxList(quandlList, idx = "Adj. Close")
}
\arguments{
\item{quandlList}{quandlList object}

\item{idx}{Character vector to index columns in quandlList}
}
\value{
List of indexed column(s) in quandList object
}
\description{
Index column(s) in quandlList object
}
