% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{logReturn}
\alias{logReturn}
\title{Calculate log returns on vector or list}
\usage{
logReturn(x, na_omit = TRUE)
}
\arguments{
\item{x}{Numeric vector or list}

\item{na_omit}{TRUE to remove NA values}
}
\value{
Vector or List of log returns on x
}
\description{
Calculate log returns on vector or list
}
