% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{quandlList}
\alias{quandlList}
\title{Fetch and store equity data from Quandl}
\usage{
quandlList(tv, start, end, src = "WIKI")
}
\arguments{
\item{tv}{Character vector of stock tickers}

\item{start}{Start date with format: "YYYY-MM-DD"}

\item{end}{End date with format: "YYYY-MM-DD"}

\item{src}{Source of Database (default "WIKI")}
}
\value{
List of equity data in xts format
}
\description{
Fetch and store equity data from Quandl
}
