% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{techList}
\alias{techList}
\title{Apply technical indicators function from TTR to
list of time series data}
\usage{
techList(tsList, FUN)
}
\arguments{
\item{tsList}{List of timeseries data with 1 column}

\item{FUN}{Vector of functions from TTR package; must only take one argument}
}
\value{
List of xts dataframes with technical indicators applied to tsList
}
\description{
Apply technical indicators function from TTR to
list of time series data
}
