% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandle.R
\name{writeCSV}
\alias{writeCSV}
\title{Write dataframe or list of dataframes to csv file(s)}
\usage{
writeCSV(x, path, replaceExisting = TRUE)
}
\arguments{
\item{x}{Dataframe or list of dataframes}

\item{path}{Path directory to save csv file(s)}

\item{replaceExisting}{TRUE to replace existing folder if it exists}
}
\description{
Write dataframe or list of dataframes to csv file(s)
}
